+++ src/main/java/hobbiedo/user/auth/Application.java
@@ -3,7 +3,9 @@ package hobbiedo.user.auth;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 @EnableDiscoveryClient
 public class Application {
