+++ src/main/java/hobbiedo/user/auth/global/api/code/status/ErrorStatus.java
@@ -15,6 +15,9 @@ public enum ErrorStatus implements BaseErrorCode {
 	USER_INTEGRATED_LOGIN_FAIL(HttpStatus.BAD_REQUEST, "USER401", "아이디 혹은 비밀번호가 일치하지 않습니다."),
 	USER_REFRESH_EXPIRED(HttpStatus.UNAUTHORIZED, "USER402", "리프레시 토큰의 유효기간이 만료되었습니다."),
 	NOT_REFRESH_TOKEN_TYPE(HttpStatus.UNAUTHORIZED, "USER403", "해당 토큰은 리프레시 토큰 타입이 아닙니다."),
+	ALREADY_USE_LOGIN_ID(HttpStatus.BAD_REQUEST, "MEMBER401", "이미 사용 중인 아이디입니다."),
+	ALREADY_USE_EMAIL(HttpStatus.BAD_REQUEST, "MEMBER402", "이미 사용 중인 이메일입니다."),
+	ALREADY_USE_PHONE_NUMBER(HttpStatus.BAD_REQUEST, "MEMBER403", "이미 사용 중인 전화번호입니다."),
 	NOT_EXIST_TOKEN(HttpStatus.NOT_FOUND, "USER404", "존재하지 않는 리프레시 토큰입니다.");
 
 	private final HttpStatus httpStatus;
@@ -32,19 +35,19 @@ public enum ErrorStatus implements BaseErrorCode {
 	@Override
 	public ErrorReasonDto getReason() {
 		return ErrorReasonDto
-				.builder()
-				.code(status)
-				.message(message)
-				.build();
+			.builder()
+			.code(status)
+			.message(message)
+			.build();
 	}
 
 	@Override
 	public ErrorReasonDto getReasonHttpStatus() {
 		return ErrorReasonDto
-				.builder()
-				.httpStatus(httpStatus)
-				.code(status)
-				.message(message)
-				.build();
+			.builder()
+			.httpStatus(httpStatus)
+			.code(status)
+			.message(message)
+			.build();
 	}
 }
