+++ src/main/java/hobbiedo/user/auth/global/api/code/status/SuccessStatus.java
@@ -11,6 +11,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum SuccessStatus implements BaseCode {
 	USER_INTEGRATED_LOGIN_SUCCESS(HttpStatus.OK, "200", "통합 로그인에 성공했습니다."),
+	INTEGRATE_SIGN_UP_SUCCESS(HttpStatus.OK, "MEMBER200", "통합 회원가입에 성공했습니다"),
 	REISSUE_TOKEN_SUCCESS(HttpStatus.OK, "200", "액세스/리프레시 토큰 재발급에 성공했습니다");
 	private final HttpStatus httpStatus;
 	private final String status;
@@ -27,19 +28,19 @@ public enum SuccessStatus implements BaseCode {
 	@Override
 	public ReasonDto getReason() {
 		return ReasonDto
-				.builder()
-				.code(status)
-				.message(message)
-				.build();
+			.builder()
+			.code(status)
+			.message(message)
+			.build();
 	}
 
 	@Override
 	public ReasonDto getReasonHttpStatus() {
 		return ReasonDto
-				.builder()
-				.httpStatus(httpStatus)
-				.code(status)
-				.message(message)
-				.build();
+			.builder()
+			.httpStatus(httpStatus)
+			.code(status)
+			.message(message)
+			.build();
 	}
 }
