+++ src/main/java/hobbiedo/user/auth/member/domain/IntegrateAuth.java
@@ -1,0 +1,36 @@
+package hobbiedo.user.auth.member.domain;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class IntegrateAuth {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	private String loginId;
+	private String password;
+
+	@OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+	@JoinColumn(name = "member_id")
+	private Member member;
+
+	@Builder
+	public IntegrateAuth(Member member, String loginId, String password) {
+		this.member = member;
+		this.loginId = loginId;
+		this.password = password;
+	}
+}
