+++ src/main/java/hobbiedo/user/auth/member/dto/request/LoginRequestDTO.java
@@ -1,0 +1,14 @@
+package hobbiedo.user.auth.member.dto.request;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class LoginRequestDTO {
+	String loginId;
+	String password;
+}
