+++ src/main/java/hobbiedo/user/auth/member/dto/request/ReIssueRequestDTO.java
@@ -1,0 +1,13 @@
+package hobbiedo.user.auth.member.dto.request;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class ReIssueRequestDTO {
+	private final String refreshToken;
+}
