+++ src/main/java/hobbiedo/user/auth/member/infrastructure/RefreshTokenRepository.java
@@ -1,0 +1,15 @@
+package hobbiedo.user.auth.member.infrastructure;
+
+import java.util.Optional;
+
+import org.springframework.data.repository.CrudRepository;
+
+import hobbiedo.user.auth.member.domain.RefreshToken;
+
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
+	Optional<RefreshToken> findById(String id);
+
+	void deleteByRefresh(String refresh);
+
+	Boolean existsByRefresh(String refresh);
+}
