+++ src/main/java/hobbiedo/user/auth/member/presentation/MemberController.java
@@ -1,0 +1,37 @@
+package hobbiedo.user.auth.member.presentation;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import hobbiedo.user.auth.global.api.ApiResponse;
+import hobbiedo.user.auth.global.api.code.status.SuccessStatus;
+import hobbiedo.user.auth.member.application.MemberService;
+import hobbiedo.user.auth.member.converter.SignUpConverter;
+import hobbiedo.user.auth.member.vo.request.IntegrateSignUpVO;
+import hobbiedo.user.auth.member.vo.response.SignUpVO;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/v1/non-users")
+@Tag(name = "Member", description = "사용자 서비스")
+public class MemberController {
+	private final MemberService memberService;
+
+	@PostMapping("/sign-up")
+	@Operation(summary = "통합 회원가입",
+		description = "통합 회원가입을 진행하여 새로운 회원을 등록합니다.")
+	public ApiResponse<SignUpVO> integrationSignUpApi(@RequestBody @Valid IntegrateSignUpVO signUpVO) {
+		return ApiResponse.onSuccess(
+			SuccessStatus.INTEGRATE_SIGN_UP_SUCCESS,
+			memberService.integrateSignUp(SignUpConverter.toDTO(signUpVO))
+		);
+	}
+}
