+++ src/main/java/hobbiedo/user/auth/member/type/GenderType.java
@@ -1,0 +1,5 @@
+package hobbiedo.user.auth.member.type;
+
+public enum GenderType {
+	FEMALE, MALE;
+}
