--- src/main/java/hobbiedo/user/auth/user/domain/Member.java
@@ -1,20 +1,0 @@
-package hobbiedo.user.auth.user.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.ToString;
-
-@Entity
-@Getter
-@ToString
-public class Member {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	private String loginId;
-	private String uuid;
-	private String password;
-}
