--- src/main/java/hobbiedo/user/auth/user/dto/request/ReIssueRequestDTO.java
@@ -1,13 +1,0 @@
-package hobbiedo.user.auth.user.dto.request;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class ReIssueRequestDTO {
-	private final String refreshToken;
-}
