+++ src/main/resources/application-local.yml
@@ -1,38 +1,57 @@
+eureka:
+  client:
+    register-with-eureka: false
+    fetch-registry: false
+    service-url:
+      defaultZone: http://172.18.0.4:8761/eureka/
+
 spring:
+  application:
+    name: auth-service
+
+  output:
+    ansi:
+      enabled: always
+
   jpa:
     database: mysql
-    hibernate.ddl-auto: update
+    hibernate.ddl-auto: create # 처음 배포할 때는 create 이후에는 update 또는 none
     show-sql: true
+
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/test?serverTimezone=Asia/Seoul
+    url: jdbc:mysql://localhost/test?serverTimezone=Asia/Seoul
     username: root
     password: 123
-  hobbie-do:
-    front-url: https://hobbie-do.site
+
   data:
     redis:
-      host: localhost
+      host: 172.18.0.3
       port: 6379
+      password: 403403
+
+hobbie-do:
+  front-url: https://hobbie-do.site
+jwt:
+  secret-key: ba223661e3ddf6a551e8a586e221069222f41c702923ce8f0b2bd313ac97c9884ae78275e3c571767d50fbfbb16f62848f4690c167ed45c81eff6af0dc1025a2
+  access-expire-time: 600000 # 10분
+  refresh-expire-time: 86400000 # 1일
 
 springdoc:
   packages-to-scan: hobbiedo.user.auth
   default-consumes-media-type: application/json;charset=UTF-8
   default-produces-media-type: application/json;charset=UTF-8
   swagger-ui:
-    path: /swagger-ui.html
+    enabled: true
+    path: /auth-service/swagger-ui.html
     disable-swagger-default-url: true
     doc-expansion: none
     tags-sorter: alpha
     operations-sorter: alpha
   api-docs:
-    path: /api-docs/json
+    enabled: true
+    path: /auth-service/v3/api-docs
     groups:
       enabled: true
   cache:
     disabled: true
-
-jwt:
-  secret-key: ba223661e3ddf6a551e8a586e221069222f41c702923ce8f0b2bd313ac97c9884ae78275e3c571767d50fbfbb16f62848f4690c167ed45c81eff6af0dc1025a2
-  access-expire-time: 60000 #10분
-  refresh-expire-time: 86400000 #1일
